queries updates

+++ server/db/schema/02_create_packages.sql
@@ -14,7 +14,8 @@ CREATE TABLE packages (
   bags_description character varying(255) NOT NULL,
   lantern_description character varying(255) NOT NULL,
   cooking_description character varying(255) NOT NULL,
-  latitude INTEGER, 
-  longitude INTEGER
+  latitude double precision, 
+  longitude double precision
  
 );
+
